plugins: extract: add new preview-extract function
It is often desirable to inspect the contents of an archive before
unpacking it.  Add a new function, preview-extract, which does this.
preview-extract is a copy-paste of extract(), but the tar commands have
parameter 'x' replaced with 't' to print the pathes of any files in the
archive.

The non-tar lines have been commented out since I rarely use those and
don't know the equivalent parameters off hand.  Hopefully a kind soul
will add those in.

Signed-off-by: Michael Turquette <mturquette@deferred.io>