Fix dklcip alias
Due to use of double-quotes ``docker ps`` was executed when the aliases
were loaded and as a consequence ``dklcip`` would always show IP of a
container that was last then instead of at runtime.

This changes the alias to use single quotes which fixes that.

It also changes the alias to use docker's ``--format`` option so
there's no need of parsing the output using grep and cut.