Added plugin for OS X Time Machine
A set of functions to simplify listing OS X Time Machine backups, and
deleting old ones. Pulled from various script solutions found on Stack
Overflow, adapted to the Bash-it framework.

The functions allow to show which backups can be found on the connected
Time Machine volume. The 'delete' function allows to safely remove all
backups except for the latest one - this can be used to free up space on
the Time Machine volume. The 'delete' function uses functionality of the
OS to safely remove the old backups, no information is lost.

The 'delete' function will only ask for the password once, then keep the credentials.
Source: http://stackoverflow.com/a/30547074/1228454