Remove alias for `brew cask search`
In Homebrew (the upstream project), `brew cask search` was merged into `brew search` a few months ago [1].

On 2018-09-30, `brew cask search` will be disabled altogether [2].

Since the output of `brew cask search` used to be subject to change at any time and was never meant to be parsed, scripts should not rely on the `bcsr` alias. So I figure it may not hurt to remove the `bcsr` alias from Bash-it before that deadline.

[1]: https://github.com/Homebrew/brew/pull/4316

[2]: https://github.com/Homebrew/brew/commit/990dd7de865ccbbc0d0ef5fb62256a86f1c92133