bash_it.sh: Require interactive shell
Previously, a fix was added for bad behavior when `bash-it` printed text
to the console in conjunction with protocols like `SCP`, as commit
83c44fac "template/profile: Require interactive shell".

It did work for those users who replace their original `~/.bash_profile`
file with the one provided by `bash-it`, but not for those who merely
append a `bash-it` call to their existing `~/.bash_profile` file.

Fix the behavior for the latter case, too.