Move external libs into vendors folder
Makefile to update vendors should maybe be a bash script. it is
however only ment to be run by developers.

Maybe vendors should be a git-subtree(1) instead Maybe we should
always supply a bash-it-init.bash(or some other name) file to the
vendoer forlder so we have a standardized way of loading vendor libs
from bash_it.sh