Avoid loops when loading libraries
The library `preexec` requires complete control over the DEBUG trap and
PROMPT_COMMAND, but it was being loaded in the library loop. When evaluating
options, I came to the conclusion that we don't have enough libraries to
justify this premature optimization, so I've flattened as much of the sourcing
as possible and added comments to explain their purpose, hoping that future
maintainers will use these to guide their decisions when adding more libraries
or changing the load order.