Tweak logic & slightly optimize git usage
This sort of test:

```sh
if CONDITION; then
	if CONDITION; then
		LIST
	fi
fi
```

The above is typically unnecessary, as the two conditions can be
combined to avoid pointless nesting.

Regarding the optimization: it's very minor, but there was an
unnecessary use of a subshell (command substitution) in testing this:

```
git rev-parse --is-inside-work-tree
```

This has been addressed, and instead of STDERR going to '/dev/null', -
now both STDOUT and STDERR do.

Although I'm pointing out this in just one instance, this sort of stuff
appears to be all over the code in this project, unfortunately.