Adding node prompt that does not require NVM
There are now alternative version managers available, such as volta.sh.
We should be able to show node version regardless of whether you are
using NVM or not. I decided not to add dedicated PREFIX variables for
now, but it can be done later.

We still check nvm first, and if it returns something — we use it.
Only if the output of NVM is blank do we use the new function to grab the version of NodeJS.

There is a caveat — if node is installed with the OS, eg `/usr/bin/node` this
function will pick it up and run `--version` on it, and therefore it will
now be displayed in the prompt.

Tested locally on OS-X/bash:

* with/without NVM
* with/without VOLTA
* with/without system node