theme/powerline-plain: remove copy of `powerline.base`
It seems like the copied functions were just slightly out of date without any altered functionality.

The one actual difference was `$POWERLINE_PROMPT_AFTER` which is blank by default and is not documented anywhere and is not used anywhere and...I will guess it's not used.

This theme seems to be about not using the powerline font, as opposed to being "simpler" in some way.