Add url() as composure metadata function
Extend composure with 'url' metadata function to support the URL fields
added to completion metadata. This allows components to link to their
official project homepages.

The url() function is defined as a no-op (like other composure metadata
functions) and is used for documentation purposes in completion files.

Fixes the undefined 'url' function issue when sourcing completions.